api/requirements.txt
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
markdown-it-py==4.0.0
mdurl==0.1.2
pycparser==2.22
Pygments==2.19.2
PyNaCl==1.5.0
rich==14.1.0
shellingham==1.5.4
typer==0.16.0
typing_extensions==4.14.1
annotated-types==0.7.0
anyio==4.10.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
fastapi==0.116.1
h11==0.16.0
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyNaCl==1.5.0
rich==14.1.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.47.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
motor==3.3.2
pymongo==4.6.1
annotated-types==0.7.0
anyio==4.10.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
fastapi==0.116.1
h11==0.16.0
idna==3.10
PyJWT==2.8.0
markdown-it-py==4.0.0
mdurl==0.1.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyNaCl==1.5.0
rich==14.1.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.47.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
motor==3.3.2
pymongo==4.6.1
